normalize the five-lines-application

+++ applications/counter/counter.c
@@ -1,3 +1,5 @@
-long counter_app(void* p) {
+#include <furi.h>
+
+int32_t counter_app(void* p) {
     return 0;
 }
